--- .babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": ["@babel/preset-env"],
-  "plugins": [
-    [
-      "@babel/plugin-transform-runtime",
-      {
-        "regenerator": true,
-        "corejs": 3
-      }
-    ]
-  ],
-  "ignore": ["source/socket.io-websocket-only.cjs"]
-}
