+++ package.json
@@ -23,25 +23,18 @@
     "main.d.ts"
   ],
   "devDependencies": {
-    "@babel/core": "^7.18.10",
-    "@babel/plugin-transform-runtime": "^7.18.10",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/register": "^7.18.9",
     "@rollup/plugin-commonjs": "^22.0.2",
-    "chai": "4.3.6",
-    "chai-as-promised": "^7.1.1",
-    "cross-env": "^7.0.3",
+    "cross-fetch": "^3.1.5",
     "eslint": "^8.21.0",
     "eslint-config-react-app": "^7.0.1",
     "isomorphic-fetch": "^3.0.0",
     "jsdom": "^20.0.0",
-    "jsdom-global": "^3.0.2",
     "lint-staged": "^13.0.3",
-    "mocha": "^10.0.0",
     "msw": "^0.47.0",
     "prettier": "^2.7.1",
     "typescript": "^4.7.4",
-    "vite": "^3.0.4"
+    "vite": "^3.0.4",
+    "vitest": "^0.23.4"
   },
   "repository": {
     "type": "git",
