+++ test/normalize_string.test.js
